Update Permission Test Suite
